refactor: HTTP 요청 객체에서 세션 관리 책임 분리
