+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Request.java
@@ -2,10 +2,6 @@ package org.apache.coyote.http11;
 
 import java.util.Collections;
 import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
-import org.apache.catalina.Manager;
-import org.apache.catalina.session.Session;
 
 public final class Http11Request {
 
@@ -15,16 +11,13 @@ public final class Http11Request {
     private final Map<String, String> headers;
     private final String body;
     private final HttpCookie httpCookie;
-    private final Manager manager;
-    private Session session;
 
     public Http11Request(
             final String method,
             final String path,
             final Map<String, String> queryParams,
             final Map<String, String> headers,
-            final String body,
-            final Manager manager
+            final String body
     ) {
         this.method = method;
         this.path = path;
@@ -32,34 +25,10 @@ public final class Http11Request {
         this.headers = headers;
         this.body = body;
         this.httpCookie = new HttpCookie(headers.get("Cookie"));
-        this.manager = manager;
     }
 
-    public static Http11Request createInvalid(final Manager manager) {
-        return new Http11Request("", "/", Collections.emptyMap(), Collections.emptyMap(), "", manager);
-    }
-
-    public Optional<Session> getSession(final boolean create) {
-        if (session != null) {
-            return Optional.of(session);
-        }
-        final Optional<Session> foundSession = httpCookie.getCookie("JSESSIONID")
-                .flatMap(manager::findSession);
-        if (foundSession.isPresent()) {
-            this.session = foundSession.get();
-            return foundSession;
-        }
-        if (create) {
-            this.session = createNewSession();
-            return Optional.of(this.session);
-        }
-        return Optional.empty();
-    }
-
-    private Session createNewSession() {
-        final var newSession = new Session(UUID.randomUUID().toString());
-        manager.add(newSession);
-        return newSession;
+    public static Http11Request createInvalid() {
+        return new Http11Request("", "/", Collections.emptyMap(), Collections.emptyMap(), "");
     }
 
     public HttpCookie getHttpCookie() {
